% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseCountsWriter.R
\name{convert10xCounts}
\alias{convert10xCounts}
\title{Convert 10x Counts to Desired Format}
\usage{
convert10xCounts(sparseCounts, format = "h5", file.name, inputPath)
}
\arguments{
\item{sparseCounts}{A sparse matrix of 10x genomics counts.}

\item{format}{A character string specifying the desired output format. Can be either "h5" or "mm".}

\item{file.name}{The name of the output file.}

\item{inputPath}{The directory path where the output file should be saved.}
}
\value{
This function does not return a value. It saves the converted counts to the specified path.
}
\description{
This function converts 10x genomics sparse count matrices to either
H5 or MM format and saves to the specified path.
}
